djreddit/requirements.txt
Django==1.11.7
factory-boy==2.9.2
ipdb==0.10.3
